--- Content/Passives/LifePassive.cs
@@ -1,13 +1,0 @@
-﻿using PathOfTerraria.Core.Systems.TreeSystem;
-
-namespace PathOfTerraria.Content.Passives;
-
-internal class LifePassive : Passive
-{
-	public override string InternalIdentifier => "AddedLife";
-	
-	public override void BuffPlayer(Player player)
-	{
-		player.statLifeMax2 += 20 * Level;
-	}
-}
