+++ Localization/en-US/Mods.PathOfTerraria.Passives.hjson
@@ -82,3 +82,8 @@ JewelSocket: {
 	Name: JewelSocket
 	Tooltip: ""
 }
+
+AddedLifeRegen: {
+	Name: LifeRegenPassive
+	Tooltip: ""
+}
